# backend/src/main/java/io/reliza/ws/tea/ProductApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.reliza.ws.tea;

import org.springframework.lang.Nullable;
import io.reliza.model.tea.TeaCle;
import io.reliza.model.tea.TeaErrorResponse;
import io.reliza.model.tea.TeaPaginatedProductReleaseResponse;
import io.reliza.model.tea.TeaProduct;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-29T10:44:15.267909500-04:00[America/Toronto]", comments = "Generator version: 7.21.0")
@Validated
@Tag(name = "CLE", description = "the CLE API")
public interface ProductApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_GET_CLE_BY_PRODUCT_ID = "/product/{uuid}/cle";
    /**
     * GET /product/{uuid}/cle
     * Get the CLE (Common Lifecycle Enumeration) data for a TEA Product
     *
     * @param uuid UUID of TEA Product in the TEA server (required)
     * @return CLE data for the requested TEA Product found and returned (status code 200)
     *         or Request was Invalid (status code 400)
     *         or Object requested by identifier not found (status code 404)
     */
    @Operation(
        operationId = "getCleByProductId",
        description = "Get the CLE (Common Lifecycle Enumeration) data for a TEA Product",
        tags = { "CLE" },
        responses = {
            @ApiResponse(responseCode = "200", description = "CLE data for the requested TEA Product found and returned", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TeaCle.class))
            }),
            @ApiResponse(responseCode = "400", description = "Request was Invalid"),
            @ApiResponse(responseCode = "404", description = "Object requested by identifier not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TeaErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "basicAuth"),
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ProductApi.PATH_GET_CLE_BY_PRODUCT_ID,
        produces = { "application/json" }
    )
    default ResponseEntity<TeaCle> getCleByProductId(
        @Parameter(name = "uuid", description = "UUID of TEA Product in the TEA server", required = true, in = ParameterIn.PATH) @PathVariable("uuid") UUID uuid
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"definitions\" : \"\", \"events\" : [ { \"eventId\" : 6, \"reason\" : \"reason\", \"references\" : [ \"https://openapi-generator.tech\", \"https://openapi-generator.tech\" ], \"identifiers\" : [ { \"idType\" : \"\", \"idValue\" : \"idValue\" }, { \"idType\" : \"\", \"idValue\" : \"idValue\" } ], \"supersededByVersion\" : \"supersededByVersion\", \"description\" : \"description\", \"published\" : \"2000-01-23T04:56:07.000+00:00\", \"type\" : \"\", \"version\" : \"version\", \"effective\" : \"2000-01-23T04:56:07.000+00:00\", \"license\" : \"license\", \"versions\" : [ { \"range\" : \"range\", \"version\" : \"version\" }, { \"range\" : \"range\", \"version\" : \"version\" } ], \"id\" : 0, \"supportId\" : \"supportId\" }, { \"eventId\" : 6, \"reason\" : \"reason\", \"references\" : [ \"https://openapi-generator.tech\", \"https://openapi-generator.tech\" ], \"identifiers\" : [ { \"idType\" : \"\", \"idValue\" : \"idValue\" }, { \"idType\" : \"\", \"idValue\" : \"idValue\" } ], \"supersededByVersion\" : \"supersededByVersion\", \"description\" : \"description\", \"published\" : \"2000-01-23T04:56:07.000+00:00\", \"type\" : \"\", \"version\" : \"version\", \"effective\" : \"2000-01-23T04:56:07.000+00:00\", \"license\" : \"license\", \"versions\" : [ { \"range\" : \"range\", \"version\" : \"version\" }, { \"range\" : \"range\", \"version\" : \"version\" } ], \"id\" : 0, \"supportId\" : \"supportId\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"OBJECT_UNKNOWN\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_RELEASES_BY_PRODUCT_ID = "/product/{uuid}/releases";
    /**
     * GET /product/{uuid}/releases
     * Get releases of the product
     *
     * @param uuid UUID of TEA Product in the TEA server (required)
     * @param pageOffset Pagination offset (optional, default to 0)
     * @param pageSize Pagination offset (optional, default to 100)
     * @return A paginated response containing TEA Product Releases (status code 200)
     *         or Request was Invalid (status code 400)
     *         or Object requested by identifier not found (status code 404)
     */
    @Operation(
        operationId = "getReleasesByProductId",
        description = "Get releases of the product",
        tags = { "TEA Product Release" },
        responses = {
            @ApiResponse(responseCode = "200", description = "A paginated response containing TEA Product Releases", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TeaPaginatedProductReleaseResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Request was Invalid"),
            @ApiResponse(responseCode = "404", description = "Object requested by identifier not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TeaErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "basicAuth"),
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ProductApi.PATH_GET_RELEASES_BY_PRODUCT_ID,
        produces = { "application/json" }
    )
    default ResponseEntity<TeaPaginatedProductReleaseResponse> getReleasesByProductId(
        @Parameter(name = "uuid", description = "UUID of TEA Product in the TEA server", required = true, in = ParameterIn.PATH) @PathVariable("uuid") UUID uuid,
        @Parameter(name = "pageOffset", description = "Pagination offset", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageOffset", required = false, defaultValue = "0") Long pageOffset,
        @Parameter(name = "pageSize", description = "Pagination offset", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageSize", required = false, defaultValue = "100") Long pageSize
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"totalResults\" : 1, \"pageStartIndex\" : 0, \"pageSize\" : 6, \"results\" : [ { \"product\" : \"\", \"components\" : [ { \"release\" : \"\", \"uuid\" : \"\" }, { \"release\" : \"\", \"uuid\" : \"\" } ], \"createdDate\" : \"\", \"releaseDate\" : \"\", \"identifiers\" : [ { \"idType\" : \"\", \"idValue\" : \"idValue\" }, { \"idType\" : \"\", \"idValue\" : \"idValue\" } ], \"uuid\" : \"\", \"version\" : \"2.24.3\", \"productName\" : \"Apache Log4j 2\", \"preRelease\" : true }, { \"product\" : \"\", \"components\" : [ { \"release\" : \"\", \"uuid\" : \"\" }, { \"release\" : \"\", \"uuid\" : \"\" } ], \"createdDate\" : \"\", \"releaseDate\" : \"\", \"identifiers\" : [ { \"idType\" : \"\", \"idValue\" : \"idValue\" }, { \"idType\" : \"\", \"idValue\" : \"idValue\" } ], \"uuid\" : \"\", \"version\" : \"2.24.3\", \"productName\" : \"Apache Log4j 2\", \"preRelease\" : true } ], \"timestamp\" : \"2024-03-20T15:30:00Z\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"OBJECT_UNKNOWN\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_TEA_PRODUCT_BY_UUID = "/product/{uuid}";
    /**
     * GET /product/{uuid}
     * Get a TEA Product by UUID
     *
     * @param uuid UUID of the TEA product in the TEA server (required)
     * @return Requested TEA Product found and returned (status code 200)
     *         or Request was Invalid (status code 400)
     *         or Object requested by identifier not found (status code 404)
     */
    @Operation(
        operationId = "getTeaProductByUuid",
        description = "Get a TEA Product by UUID",
        tags = { "TEA Product" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Requested TEA Product found and returned", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TeaProduct.class))
            }),
            @ApiResponse(responseCode = "400", description = "Request was Invalid"),
            @ApiResponse(responseCode = "404", description = "Object requested by identifier not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TeaErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "basicAuth"),
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ProductApi.PATH_GET_TEA_PRODUCT_BY_UUID,
        produces = { "application/json" }
    )
    default ResponseEntity<TeaProduct> getTeaProductByUuid(
        @Parameter(name = "uuid", description = "UUID of the TEA product in the TEA server", required = true, in = ParameterIn.PATH) @PathVariable("uuid") UUID uuid
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"identifiers\" : [ { \"idType\" : \"\", \"idValue\" : \"idValue\" }, { \"idType\" : \"\", \"idValue\" : \"idValue\" } ], \"name\" : \"name\", \"uuid\" : \"\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"OBJECT_UNKNOWN\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}