# tea-spec/openapi.yaml
...
openapi: 3.1.1
...
info:
  title: Transparency Exchange API
  summary: The OWASP Transparency Exchange API specification for consumers and publishers
  description: TBC
  contact:
    name: TEA Working Group
    email: tbc@somewhere.tld
    url: https://github.com/CycloneDX/transparency-exchange-api
  license:
    name: Apache 2.0
    url: https://github.com/CycloneDX/transparency-exchange-api/blob/main/LICENSE
  version: 0.4.0
servers:
  - url: http://localhost/tea/v1
    description: Local development
paths:
  /product/{uuid}:
    get:
      description: Get a TEA Product by UUID
      operationId: getTeaProductByUuid
      parameters:
        - name: uuid
          in: path
          required: true
          description: UUID of the TEA product in the TEA server
          schema:
            "$ref": "#/components/schemas/uuid"
      responses:
        '200':
          description: Requested TEA Product found and returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/product"
        '400':
          $ref: "#/components/responses/400-invalid-request"
        '404':
          $ref: "#/components/responses/404-object-by-id-not-found"
      tags:
        - TEA Product
  /product/{uuid}/releases:
    get:
      description: Get releases of the product
      operationId: getReleasesByProductId
      parameters:
        - name: uuid
          in: path
          required: true
          description: UUID of TEA Product in the TEA server
          schema:
            "$ref": "#/components/schemas/uuid"
        - $ref: "#/components/parameters/page-offset"
        - $ref: "#/components/parameters/page-size"
      responses:
        '200':
          $ref: "#/components/responses/paginated-product-release"
        '400':
          $ref: "#/components/responses/400-invalid-request"
        '404':
          $ref: "#/components/responses/404-object-by-id-not-found"
      tags:
        - TEA Product Release
...
  /product/{uuid}/cle:
    get:
      description: Get the CLE (Common Lifecycle Enumeration) data for a TEA Product
      operationId: getCleByProductId
      parameters:
        - name: uuid
          in: path
          required: true
          description: UUID of TEA Product in the TEA server
          schema:
            "$ref": "#/components/schemas/uuid"
      responses:
        '200':
          description: CLE data for the requested TEA Product found and returned
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/cle"
        '400':
          $ref: "#/components/responses/400-invalid-request"
        '404':
          $ref: "#/components/responses/404-object-by-id-not-found"
      tags:
        - CLE
...
components:
  schemas:
...
    date-time:
      type: string
      description: Timestamp
      format: date-time
      pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"
      example: '2024-03-20T15:30:00Z'
    identifier:
      type: object
      description: An identifier with a specified type
      properties:
        idType:
          description: Type of identifier, e.g. `TEI`, `PURL`, `CPE`
          "$ref": "#/components/schemas/identifier-type"
        idValue:
          description: Identifier value
          type: string
    identifier-type:
      type: string
      description: Enumeration of identifiers types
      enum:
        - CPE
        - TEI
        - PURL
        - COMPLIANCE_DOCUMENT
...
    uuid:
      type: string
      description: A UUID
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"

    #
    # TEA Product
    #
    product:
      type: object
      description: A TEA product
      properties:
        uuid:
          description: A unique identifier for the TEA product
          "$ref": "#/components/schemas/uuid"
        name:
          type: string
          description: Product name
        identifiers:
          type: array
          description: |
            List of identifiers for the product, like TEI, CPE, PURL or other identifiers
          items:
            "$ref": "#/components/schemas/identifier"
      required:
        - uuid
        - name
        - identifiers
      examples:
        - uuid: 09e8c73b-ac45-4475-acac-33e6a7314e6d
          name: Apache Log4j 2
          identifiers:
            - idType: CPE
              idValue: cpe:2.3:a:apache:log4j
            - idType: PURL
              idValue: pkg:maven/org.apache.logging.log4j/log4j-api
    #
    # TEA Product Release
    #
    productRelease:
      type: object
      description: A specific release of a TEA product
      properties:
        uuid:
          description: A unique identifier for the TEA Product Release
          "$ref": "#/components/schemas/uuid"
        product:
          description: UUID of the TEA Product this release belongs to
          "$ref": "#/components/schemas/uuid"
        productName:
          description: Name of the TEA Product this release belongs to
          type: string
          example: Apache Log4j 2
        version:
          description: Version number of the product release
          type: string
          example: 2.24.3
        createdDate:
          description: Timestamp when this Product Release was created in TEA (for sorting purposes)
          "$ref": "#/components/schemas/date-time"
        releaseDate:
          description: Timestamp of the product release
          "$ref": "#/components/schemas/date-time"
        preRelease:
          type: boolean
          description: |
            A flag indicating pre-release (or beta) status.
            May be disabled after the creation of the release object, but can't be enabled after creation of an object.
        identifiers:
          type: array
          description: List of identifiers for the product release
          items:
            "$ref": "#/components/schemas/identifier"
        components:
          type: array
          description: |
            List of component references that compose this product release. A component reference can optionally include
            the UUID of a specific component release to pin the exact version.
          items:
            "$ref": "#/components/schemas/component-ref"
      required:
        - uuid
        - version
        - createdDate
        - components
      examples:
        - uuid: 123e4567-e89b-12d3-a456-426614174000
          version: "2.24.3"
          createdDate: 2025-04-01T15:43:00Z
          releaseDate: 2025-04-01T15:43:00Z
          identifiers:
            - idType: TEI
              idValue: tei:vendor:product@2.24.3
          components:
            - uuid: 3910e0fd-aff4-48d6-b75f-8bf6b84687f0
            - uuid: b844c9bd-55d6-478c-af59-954a932b6ad3
              release: da89e38e-95e7-44ca-aa7d-f3b6b34c7fab

    #
    # TEA Component and related objects
    #
...
    component-ref:
      type: object
      description: A reference to a TEA component or specific component release
      properties:
        uuid:
          description: A unique identifier for the TEA component
          "$ref": "#/components/schemas/uuid"
        release:
          description: |
            Optional UUID of a specific release included in the product in the case where the product
            always include a specific release of a component. The product name should include a version
            identifier in this case.
          "$ref": "#/components/schemas/uuid"
      required:
        - uuid



    #
    # TEA Component Release and related objects
    #
...
    unknown-error-type:
      type: string
      description: Classification of TEA error response
      enum:
        - OBJECT_UNKNOWN
        - OBJECT_NOT_SHAREABLE
    #
    # Types used in API responses
    #
    error-response:
      type: object
      description: Error response
      additionalProperties: false
      properties:
        error:
          $ref: "#/components/schemas/unknown-error-type"
      required:
        - error
...
    cle-event-type:
      type: string
      description: The type of CLE lifecycle event
      enum:
        - released
        - endOfDevelopment
        - endOfSupport
        - endOfLife
        - endOfDistribution
        - endOfMarketing
        - supersededBy
        - componentRenamed
        - withdrawn
    cle-version-specifier:
      type: object
      description: A version specifier that can be either a single version or a version range
      properties:
        version:
          type: string
          description: A specific version string
        range:
          type: string
          description: A version range in vers format (e.g. "vers:npm/>=1.0.0|<2.0.0")
    cle-event:
      type: object
      description: A discrete lifecycle event from the CLE specification
      properties:
        id:
          type: integer
          description: A unique, auto-incrementing integer identifier for the event
        type:
          description: The type of lifecycle event
          $ref: "#/components/schemas/cle-event-type"
        effective:
          type: string
          format: date-time
          description: ISO 8601 timestamp (UTC) when the event takes effect
        published:
          type: string
          format: date-time
          description: ISO 8601 timestamp (UTC) when the event was first published
        version:
          type: string
          description: Version string (used by released event type)
        versions:
          type: array
          description: List of version specifiers affected by this event
          items:
            $ref: "#/components/schemas/cle-version-specifier"
        supportId:
          type: string
          description: Reference to a support policy ID defined in the definitions section
        license:
          type: string
          description: License identifier (used by released event type)
        supersededByVersion:
          type: string
          description: Version string that supersedes the affected versions (used by supersededBy event type)
        identifiers:
          type: array
          description: New identifiers for the component (used by componentRenamed event type)
          items:
            $ref: "#/components/schemas/identifier"
        eventId:
          type: integer
          description: ID of the event being withdrawn (used by withdrawn event type)
        reason:
          type: string
          description: Human-readable explanation (used by withdrawn event type)
        description:
          type: string
          description: Human-readable description of the event
        references:
          type: array
          description: List of URLs to supporting documentation
          items:
            type: string
            format: uri
      required:
        - id
        - type
        - effective
        - published
      examples:
        - id: 1
          type: released
          effective: "2024-01-01T00:00:00Z"
          version: "1.0.0"
          license: MIT
          published: "2023-06-01T00:00:00Z"
        - id: 3
          type: endOfSupport
          effective: "2024-01-01T00:00:00Z"
          versions:
            - version: "1.0.0"
          supportId: standard
          published: "2023-06-01T00:00:00Z"
    cle-support-definition:
      type: object
      description: A support policy definition from CLE
      properties:
        id:
          type: string
          description: Unique identifier for the support policy
        description:
          type: string
          description: Human-readable description of the policy
        url:
          type: string
          format: uri
          description: URL to detailed documentation about this support policy
      required:
        - id
        - description
      examples:
        - id: standard
          description: Standard product support policy
          url: https://example.com/support/standard
    cle-definitions:
      type: object
      description: Container for reusable CLE policy definitions
      properties:
        support:
          type: array
          description: List of support policies
          items:
            $ref: "#/components/schemas/cle-support-definition"
    cle:
      type: object
      description: |
        Common Lifecycle Enumeration (CLE) object based on ECMA-428 TC54 TG3 CLE Specification v1.0.0.
        Contains lifecycle events and optional reusable definitions for a component or product.
      properties:
        events:
          type: array
          description: |
            Ordered array of CLE Event objects representing lifecycle events.
            MUST be ordered by ID in descending order (newest events with highest IDs first).
          items:
            $ref: "#/components/schemas/cle-event"
        definitions:
          description: Container for reusable policy definitions referenced by events
          $ref: "#/components/schemas/cle-definitions"
      required:
        - events
      examples:
        - events:
            - id: 3
              type: endOfSupport
              effective: "2025-06-01T00:00:00Z"
              versions:
                - range: "vers:npm/>=1.0.0|<2.0.0"
              supportId: standard
              published: "2025-01-01T00:00:00Z"
            - id: 2
              type: endOfDevelopment
              effective: "2025-01-01T00:00:00Z"
              versions:
                - version: "1.0.0"
              supportId: standard
              published: "2024-06-01T00:00:00Z"
            - id: 1
              type: released
              effective: "2024-01-01T00:00:00Z"
              version: "1.0.0"
              license: Apache-2.0
              published: "2024-01-01T00:00:00Z"
          definitions:
            support:
              - id: standard
                description: Standard product support policy
                url: https://example.com/support/standard
    pagination-details:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2024-03-20T15:30:00Z'
        pageStartIndex:
          type: integer
          format: int64
          default: 0
        pageSize:
          type: integer
          format: int64
          default: 100
        totalResults:
          type: integer
          format: int64
      required:
        - timestamp
        - pageStartIndex
        - pageSize
        - totalResults
...
    paginated-product-release-response:
      type: object
      description: A paginated response containing TEA Product Releases
      allOf:
        - $ref: "#/components/schemas/pagination-details"
        - type: object
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/productRelease"
...
  responses:
...
    400-invalid-request:
      description: Request was Invalid
      content:
        application/json: {}
...
    404-object-by-id-not-found:
      description: Object requested by identifier not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error-response"
...
    paginated-product-release:
      description: A paginated response containing TEA Product Releases
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/paginated-product-release-response"
...
  parameters:
...
    page-offset:
      name: pageOffset
      description: Pagination offset
      in: query
      required: false
      schema:
        type: integer
        format: int64
        default: 0
    page-size:
      name: pageSize
      description: Pagination offset
      in: query
      required: false
      schema:
        type: integer
        format: int64
        default: 100
    #
    # Query by identifier
    #
    # Since OpenAPI 3.0 it is possible to use RFC 6570-based serialization for JSON parameters of type array or object:
    #   https://swagger.io/docs/specification/v3_0/serialization/
    #
    # Unfortunately many tools don't support it, for example,
    # the `openapi-generator` for Java does not handle this correctly.
    #   https://github.com/OpenAPITools/openapi-generator/issues/4808
    #
    # This can be uncommented, when RFC 6570-base serialization reaches a wider adoption:
    #
    # identifier-param:
    #   name: identifierParam
    #   description: If present, only the objects with the given identifier will be returned.
    #   in: query
    #   schema:
    #     $ref: "#/components/schemas/identifier"
    #   style: form
    #   explode: true
    #
    # In the meantime we explode the object manually:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    basicAuth:
      type: http
      scheme: basic
security:
  - bearerAuth: []
  - basicAuth: []
...